2023.05.17

+++ src/main/java/com/matsinger/barofishserver/user/UserRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findById(Long id);
+public interface UserRepository extends JpaRepository<User, Integer> {
+    Optional<User> findById(Integer id);
 }
